+++ CHANGELOG.md
@@ -15,7 +15,21 @@
 ### Added
 ### Deprecated
 ### Removed
+### Changed
+
+* Mouse cursor is now always a `left_ptr` when inside the margins, to
+  indicate it is not possible to start a selection.
+* Scrollback position indicator. This feature is optional and
+  controlled by the **scrollback-indicator-style** and
+  **scrollback-indicator-format** options in `footrc`
+  (https://codeberg.org/dnkl/foot/issues/42).
+
+
 ### Fixed
+
+* Crash when starting a selection inside the margins.
+
+
 ### Security
 ### Contributors